Added two new flags for downloads.
The first allows downloads to be flagged as one which shouldn't be considered by move-on-completion and move-on-removal rules.

The second is used to protect against attempting to move a download more than once for move-on-completion - it can get invoked more than once if a completed download is stopped and then started again.